+++ cli/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 2.5.0
 - Remove `--hint` argument from CLI.
 - Change behaviour of `--origin` such that it overwrites the internal file `static_cuas_location` member.
-- Add `--ignore-cuas-origin` option
+- Add `--ignore-cuas-origin` option.
+- Add `-o` option to specify output path.
 
 ## 2.4.0
 - Separate track records by classification: Create one KML track per set of records with the same classification in a COURAGEOUS track.